# specs/001-cimientos-clientes/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Joyería Blanco — API de gestión
  version: 0.1.0
  description: |
    Contrato de la feature 001. La web y la API comparten origen, así que no hay CORS.

    **Autenticación**: cookie de sesión `__Host-jb_sesion` (`jb_sesion` en desarrollo),
    `HttpOnly; Secure; SameSite=Strict`.

    **Peticiones que modifican datos** (POST, PUT, PATCH, DELETE):
    - Exigen la cabecera `X-CSRF-Token` con el token de la sesión (excepto `POST /sesion`).
    - Exigen un `Origin` permitido, también en `POST /sesion`.

    **Errores**: RFC 9457 `application/problem+json`, con mensajes en español.

    **Contraseña temporal**: si la sesión la tiene, solo se permiten `GET /sesion`,
    `DELETE /sesion` y `PUT /cuenta/contrasena`. El resto responde 403 con
    `type: .../contrasena-temporal`.

    **Catálogo de `type` de error** (`/problemas/...`):

    | `type` | Estado HTTP |
    |---|---|
    | `validacion` | 422, con `errores` por campo |
    | `no-autenticado` | 401 |
    | `credenciales` | 401, mensaje genérico (FR-007) |
    | `sin-permiso` | 403 |
    | `csrf` | 403 |
    | `origen` | 403 |
    | `contrasena-temporal` | 403 |
    | `no-encontrado` | 404 |
    | `conflicto-version` | 409 |
    | `duplicado` | 409, con `cliente_existente` |
    | `ultimo-administrador` | 409 |
    | `autogestion` | 409, no se permite sobre la propia cuenta |
    | `cliente-con-documentos` | 409 |
...
    | `limite-origen` | 429 |
    | `interno` | 500, sin detalles |

...

...
servers:
...
paths:
...
  /v1/sesion:
    post:
      tags: [sesion]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredencialesEntrada" }
      responses:
        "200":
          description: Sesión creada. Establece la cookie de sesión.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SesionSalida" }
        "401":
          description: >
            Mensaje genérico idéntico para usuario inexistente, contraseña incorrecta,
            cuenta bloqueada o desactivada (FR-007).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problema" }
        "403": { description: Origen no permitido, $ref: "#/components/responses/Problema" }
        "429": { description: Límite de intentos por origen superado (FR-006), $ref: "#/components/responses/Problema" }
        "422": { $ref: "#/components/responses/Validacion" }
    get:
      tags: [sesion]
      summary: Sesión actual (usuario, rol, token CSRF y caducidad)
      responses:
        "200":
          description: Sesión vigente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SesionSalida" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
    delete:
      tags: [sesion]
      summary: Cerrar sesión (invalida la sesión en el servidor)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "204": { description: Sesión cerrada; la cookie se borra }
        "401": { $ref: "#/components/responses/NoAutenticado" }
  /v1/cuenta/contrasena:
    put:
      tags: [cuenta]
      summary: Cambiar la propia contraseña (FR-019). Revoca las demás sesiones.
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CambioContrasenaEntrada" }
      responses:
        "204": { description: Contraseña cambiada }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "422": { $ref: "#/components/responses/Validacion" }
  /v1/usuarios:
    get:
      tags: [usuarios]
      summary: Listar usuarios (solo administradores)
      description: |
        Por defecto excluye a los usuarios eliminados (FR-061). Con `incluir_eliminados=true`
        también los devuelve, marcados con `eliminado`, para el filtro de la auditoría.
      parameters:
        - name: incluir_eliminados
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Usuarios
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UsuarioSalida" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "403": { $ref: "#/components/responses/Prohibido" }
    post:
      tags: [usuarios]
      summary: Alta de usuario con contraseña temporal generada (FR-014, FR-015)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsuarioAltaEntrada" }
      responses:
        "201":
          description: Usuario creado. La contraseña temporal se devuelve SOLO en esta respuesta.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsuarioConContrasenaTemporalSalida" }
        "403": { $ref: "#/components/responses/Prohibido" }
        "409": { description: Nombre de usuario ya existente, $ref: "#/components/responses/Problema" }
        "422": { $ref: "#/components/responses/Validacion" }
  /v1/usuarios/{usuarioId}:
    parameters: [{ $ref: "#/components/parameters/UsuarioId" }]
    description: |
      Un usuario eliminado se trata como inexistente en todas las operaciones de gestión (404).
    get:
      tags: [usuarios]
      summary: Detalle de usuario (solo administradores)
      responses:
        "200":
          description: Usuario
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsuarioSalida" }
        "403": { $ref: "#/components/responses/Prohibido" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
    patch:
      tags: [usuarios]
      summary: Cambiar nombre visible o rol (regla del último administrador, FR-017)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsuarioEdicionEntrada" }
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsuarioSalida" }
        "403": { $ref: "#/components/responses/Prohibido" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
        "409": { description: Dejaría el sistema sin administrador activo, $ref: "#/components/responses/Problema" }
        "422": { $ref: "#/components/responses/Validacion" }
    delete:
      tags: [usuarios]
      summary: Eliminar un usuario desactivado (FR-061)
      description: |
        Irreversible. El usuario desaparece de la gestión, no puede volver a identificarse y su
        nombre de usuario queda libre. Se descartan su contraseña y sus sesiones. Lo que registró
        (clientes, auditoría) se conserva sin cambios y sus referencias pasan a llevar
        `eliminado: true`. Queda el evento `usuario_eliminado` en la auditoría.
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "204": { description: Usuario eliminado }
        "403": { $ref: "#/components/responses/Prohibido" }
        "404": { description: No existe o ya está eliminado, $ref: "#/components/responses/NoEncontrado" }
        "409": { description: "`usuario-activo` (hay que desactivarlo antes) o `autogestion`", $ref: "#/components/responses/Problema" }
  /v1/usuarios/{usuarioId}/desactivacion:
    parameters: [{ $ref: "#/components/parameters/UsuarioId" }]
    post:
      tags: [usuarios]
      summary: Desactivar usuario y revocar todas sus sesiones (FR-016)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "200":
          description: Usuario desactivado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsuarioSalida" }
        "403": { $ref: "#/components/responses/Prohibido" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
        "409": { description: Último administrador activo o autodesactivación, $ref: "#/components/responses/Problema" }
  /v1/usuarios/{usuarioId}/reactivacion:
    parameters: [{ $ref: "#/components/parameters/UsuarioId" }]
    post:
      tags: [usuarios]
      summary: Reactivar usuario
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "200":
          description: Usuario reactivado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsuarioSalida" }
        "403": { $ref: "#/components/responses/Prohibido" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
  /v1/usuarios/{usuarioId}/restablecimiento-contrasena:
    parameters: [{ $ref: "#/components/parameters/UsuarioId" }]
    post:
      tags: [usuarios]
      summary: Restablecer contraseña; levanta el bloqueo y revoca las sesiones (FR-015, FR-016)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "200":
          description: Contraseña temporal nueva, devuelta SOLO en esta respuesta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsuarioConContrasenaTemporalSalida" }
        "403": { $ref: "#/components/responses/Prohibido" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
  /v1/clientes:
    get:
      tags: [clientes]
      summary: Listado paginado con búsqueda, filtros y orden (FR-031 a FR-033)
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 100 }, description: "Búsqueda parcial en nombre, identificación y localidad, sin tildes ni mayúsculas" }
        - { name: provincia, in: query, schema: { type: string, pattern: "^[0-9]{2}$" } }
        - { name: tipo, in: query, schema: { $ref: "#/components/schemas/TipoCliente" } }
        - { name: estado, in: query, schema: { type: string, enum: [activos, inactivos, todos], default: activos } }
        - { name: orden, in: query, schema: { type: string, enum: [nombre_asc, nombre_desc, recientes, antiguos], default: nombre_asc } }
        - { $ref: "#/components/parameters/Pagina" }
        - { $ref: "#/components/parameters/Tamano" }
      responses:
        "200":
          description: Página de clientes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginaClientes" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "422": { $ref: "#/components/responses/Validacion" }
    post:
      tags: [clientes]
      summary: Alta de cliente (FR-023 a FR-029)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClienteEntrada" }
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClienteSalida" }
        "409":
          description: Identificación ya registrada. `cliente_existente` indica cuál y si está inactivo.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemaDuplicado" }
        "422": { $ref: "#/components/responses/Validacion" }
...
  /v1/clientes/{clienteId}:
    parameters: [{ $ref: "#/components/parameters/ClienteId" }]
    get:
      tags: [clientes]
      summary: Ficha del cliente con trazabilidad
      responses:
        "200":
          description: Cliente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClienteSalida" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
    put:
      tags: [clientes]
      summary: Editar cliente con concurrencia optimista (FR-030)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClienteEdicionEntrada" }
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClienteSalida" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
        "409": { description: "Versión desfasada (el cliente ha cambiado) o identificación duplicada", $ref: "#/components/responses/Problema" }
        "422": { $ref: "#/components/responses/Validacion" }
    delete:
      tags: [clientes]
      summary: Borrado definitivo (solo administradores y solo sin documentos, FR-037)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "204": { description: Cliente borrado (queda instantánea en auditoría) }
        "403": { $ref: "#/components/responses/Prohibido" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
        "409": { description: El cliente tiene documentos; hay que desactivarlo, $ref: "#/components/responses/Problema" }
  /v1/clientes/{clienteId}/desactivacion:
    parameters: [{ $ref: "#/components/parameters/ClienteId" }]
    post:
      tags: [clientes]
      summary: Desactivar cliente (FR-036)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "200":
          description: Cliente desactivado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClienteSalida" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
  /v1/clientes/{clienteId}/reactivacion:
    parameters: [{ $ref: "#/components/parameters/ClienteId" }]
    post:
      tags: [clientes]
      summary: Reactivar cliente (FR-036)
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      responses:
        "200":
          description: Cliente reactivado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClienteSalida" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
components:
  securitySchemes:
    cookieSesion:
      type: apiKey
      in: cookie
      name: __Host-jb_sesion
  parameters:
    Csrf:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    UsuarioId:
      name: usuarioId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ClienteId:
      name: clienteId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Pagina:
      name: pagina
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Tamano:
      name: tamano
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
  responses:
    Problema:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problema" }
    Validacion:
      description: Datos no válidos (errores por campo)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problema" }
    NoAutenticado:
      description: Sin sesión válida (inexistente, caducada o revocada)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problema" }
    Prohibido:
      description: Sin permiso, CSRF u origen no válido, o contraseña temporal pendiente de cambio
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problema" }
    NoEncontrado:
      description: Recurso inexistente
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problema" }
  schemas:
    Problema:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference, example: "/problemas/validacion" }
        title: { type: string, example: "Datos no válidos" }
        status: { type: integer }
        detail: { type: string }
        errores:
          type: array
          items:
            type: object
            required: [campo, mensaje]
            properties:
              campo: { type: string, example: "identificacion_numero" }
              mensaje: { type: string, example: "El carácter de control del NIF no es correcto" }
    ProblemaDuplicado:
      allOf:
        - $ref: "#/components/schemas/Problema"
        - type: object
          required: [cliente_existente]
          properties:
            cliente_existente:
              type: object
              required: [id, nombre, activo]
              properties:
                id: { type: string, format: uuid }
                nombre: { type: string }
                activo: { type: boolean }
    Rol: { type: string, enum: [administrador, empleado] }
    TipoCliente: { type: string, enum: [particular, empresa] }
    TipoIdentificacion:
      type: string
      description: >
        `NIF` (solo país ES) o clave de la lista L7 de DsRegistroVeriFactu.xlsx v1.0 (F-2). Con país
        ES solo se admiten NIF o 03. El 02 solo se admite para los Estados de la tabla NIF-IVA
        (Validaciones v1.2.2, nota 1). El 07 se decide al generar el registro de facturación
        (research R-20.1).
      enum: [NIF, "02", "03", "04", "05", "06"]
...
    CredencialesEntrada:
      type: object
      additionalProperties: false
      required: [nombre_usuario, contrasena]
      properties:
        nombre_usuario: { type: string, minLength: 1, maxLength: 50 }
        contrasena: { type: string, minLength: 1, maxLength: 128 }
    SesionSalida:
      type: object
      required: [usuario, csrf_token, expira_en, inactividad_segundos]
      properties:
        usuario: { $ref: "#/components/schemas/UsuarioSalida" }
        csrf_token: { type: string }
        expira_en: { type: string, format: date-time, description: Caducidad absoluta }
        inactividad_segundos: { type: integer, description: Ventana de inactividad configurada }
    CambioContrasenaEntrada:
      type: object
      additionalProperties: false
      required: [contrasena_actual, contrasena_nueva]
      properties:
        contrasena_actual: { type: string, minLength: 1, maxLength: 128 }
        contrasena_nueva: { type: string, minLength: 12, maxLength: 128 }
    UsuarioSalida:
      type: object
      required: [id, nombre_usuario, nombre, rol, activo, eliminado, contrasena_temporal, bloqueado, creado_en]
      properties:
        id: { type: string, format: uuid }
        nombre_usuario: { type: string }
        nombre: { type: string }
        rol: { $ref: "#/components/schemas/Rol" }
        activo: { type: boolean }
        eliminado: { type: boolean, description: "Solo aparece a true con incluir_eliminados (FR-061)" }
        contrasena_temporal: { type: boolean }
        bloqueado: { type: boolean, description: Bloqueo temporal vigente }
        ultimo_acceso_en: { type: [string, "null"], format: date-time }
        creado_en: { type: string, format: date-time }
    UsuarioAltaEntrada:
      type: object
      additionalProperties: false
      required: [nombre_usuario, nombre, rol]
      properties:
        nombre_usuario: { type: string, pattern: "^[a-zA-Z0-9._-]{3,50}$" }
        nombre: { type: string, minLength: 1, maxLength: 120 }
        rol: { $ref: "#/components/schemas/Rol" }
    UsuarioEdicionEntrada:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nombre: { type: string, minLength: 1, maxLength: 120 }
        rol: { $ref: "#/components/schemas/Rol" }
    UsuarioConContrasenaTemporalSalida:
      type: object
      required: [usuario, contrasena_temporal]
      properties:
        usuario: { $ref: "#/components/schemas/UsuarioSalida" }
        contrasena_temporal: { type: string, description: Se muestra una única vez }
    ClienteEntrada:
      type: object
      additionalProperties: false
      required: [tipo, nombre, identificacion_pais, identificacion_tipo, identificacion_numero]
      properties:
        tipo: { $ref: "#/components/schemas/TipoCliente" }
        nombre: { type: string, minLength: 1, maxLength: 120, description: "NombreRazon (F-1)" }
        identificacion_pais: { type: string, pattern: "^[A-Za-z]{2}$", default: ES }
        identificacion_tipo: { $ref: "#/components/schemas/TipoIdentificacion" }
        identificacion_numero: { type: string, minLength: 1, maxLength: 30, description: "Se normaliza a ≤20 caracteres (NIF: 9)" }
        direccion: { type: [string, "null"], maxLength: 200 }
        codigo_postal: { type: [string, "null"], maxLength: 10 }
        localidad: { type: [string, "null"], maxLength: 100 }
        provincia_codigo: { type: [string, "null"], pattern: "^[0-9]{2}$", description: "Si país=ES y hay CP, el servidor la deriva del CP" }
        provincia_texto: { type: [string, "null"], maxLength: 100, description: Solo si el país de residencia no es España }
        pais_residencia: { type: string, pattern: "^[A-Za-z]{2}$", default: ES }
        telefono: { type: [string, "null"], maxLength: 30 }
        correo: { type: [string, "null"], format: email, maxLength: 254 }
        observaciones: { type: [string, "null"], maxLength: 2000 }
    ClienteEdicionEntrada:
      allOf:
        - $ref: "#/components/schemas/ClienteEntrada"
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 1, description: Versión leída; si no coincide, 409 }
    ClienteResumenSalida:
      type: object
      required: [id, tipo, nombre, identificacion_pais, identificacion_tipo, identificacion_numero, activo]
      properties:
        id: { type: string, format: uuid }
        tipo: { $ref: "#/components/schemas/TipoCliente" }
        nombre: { type: string }
        identificacion_pais: { type: string }
        identificacion_tipo: { $ref: "#/components/schemas/TipoIdentificacion" }
        identificacion_numero: { type: string }
        localidad: { type: [string, "null"] }
        provincia_nombre: { type: [string, "null"], description: "nombre_visible de la provincia INE o texto libre" }
        telefono: { type: [string, "null"] }
        correo: { type: [string, "null"] }
        activo: { type: boolean }
    ClienteSalida:
      allOf:
        - $ref: "#/components/schemas/ClienteResumenSalida"
        - type: object
          required: [pais_residencia, version, creado_en, creado_por, actualizado_en, actualizado_por]
          properties:
            direccion: { type: [string, "null"] }
            codigo_postal: { type: [string, "null"] }
            provincia_codigo: { type: [string, "null"] }
            provincia_texto: { type: [string, "null"] }
            pais_residencia: { type: string }
            observaciones: { type: [string, "null"] }
            version: { type: integer }
            creado_en: { type: string, format: date-time }
            creado_por: { $ref: "#/components/schemas/UsuarioReferencia" }
            actualizado_en: { type: string, format: date-time }
            actualizado_por: { $ref: "#/components/schemas/UsuarioReferencia" }
    UsuarioReferencia:
      type: object
      required: [id, nombre, eliminado]
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        eliminado: { type: boolean, description: "La web muestra el nombre con la marca «(eliminado)» (FR-061)" }
    PaginaClientes:
      type: object
      required: [elementos, total, pagina, tamano]
      properties:
        elementos:
          type: array
          items: { $ref: "#/components/schemas/ClienteResumenSalida" }
        total: { type: integer }
        pagina: { type: integer }
        tamano: { type: integer }
...
security:
  - cookieSesion: []